+++ src/pdm_conda/__init__.py
@@ -18,4 +18,4 @@ def main(core: Core):
         core.add_config(name, config)
 
 
-__version__ = "0.8.0"
+__version__ = "0.8.1"