+++ src/pdm_conda/installers/synchronizers.py
@@ -5,6 +5,7 @@ from pdm.models.candidates import Candidate
 
 from pdm_conda.models.candidates import CondaCandidate
 from pdm_conda.models.environment import CondaEnvironment, Environment
+from pdm_conda.models.setup import CondaSetupDistribution
 
 
 class CondaSynchronizer(Synchronizer):
@@ -44,16 +45,20 @@ class CondaSynchronizer(Synchronizer):
         if to_remove:
             to_remove = [p for p in to_remove if p not in self.environment.python_dependencies]
 
-        if self.parallel and any(True for d in to_remove + to_update if isinstance(self.candidates[d], CondaCandidate)):
+        num_update, num_remove = (
+            len([p for p in pks if isinstance(self.working_set[p], CondaSetupDistribution)])
+            for pks in (to_update, to_remove)
+        )
+        if self.parallel and (num_update + num_remove > 0):
             self.environment.project.core.ui.echo("Deactivating parallel uninstall.")
             self.parallel = False
 
         if self.environment.project.conda_config.batched_commands:
+            num_adds = len([p for p in to_add if isinstance(self.candidates[p], CondaCandidate)])
+
             self.manager.prepare_batch_operations(
-                *(
-                    len([p for p in pkgs if isinstance(self.candidates[p], CondaCandidate)])
-                    for pkgs in (to_add + to_update, to_remove + to_update)
-                )
+                num_install=num_adds + num_update,
+                num_remove=num_remove + num_update,
             )
 
         return to_add, to_update, to_remove