+++ tests/test_list.py
@@ -3,12 +3,11 @@ import pytest
 from tests.conftest import CONDA_INFO, CONDA_MAPPING
 
 
-@pytest.mark.parametrize("empty_conda_list", [False])
 class TestList:
-    @pytest.mark.parametrize("conda_response", CONDA_INFO)
+    @pytest.mark.parametrize("conda_info", CONDA_INFO)
     @pytest.mark.parametrize("conda_mapping", CONDA_MAPPING)
     @pytest.mark.parametrize("runner", ["micromamba"])
-    def test_list(self, pdm, project, conda, conda_response, runner, mock_conda_mapping):
+    def test_list(self, pdm, project, conda, conda_info, runner, mock_conda_mapping, installed_packages, working_set):
         """
         Test `list` command work as expected
         """
@@ -19,16 +18,21 @@ class TestList:
                     "pdm": {
                         "conda": {
                             "runner": runner,
-                            "dependencies": ["dep-pip"],
                         },
                     },
                 },
             },
         )
+        # fake installation
+        for p in conda_info:
+            if p["name"] not in [ip["name"] for ip in installed_packages]:
+                installed_packages.append(p)
         result = pdm(["list"], obj=project)
         assert result.exception is None
 
-        for p in conda_response:
+        for p in conda_info:
+            assert p["name"] in result.stdout
+        for p in installed_packages:
             assert p["name"] in result.stdout
         for (cmd,), _ in conda.call_args_list:
             assert cmd[0] == runner